Improved argument handling
Some services can have spaces due to pathname and/or because of arguments after the executable (e.g.: c:\program files\neo4j\bin\prunsrv-amd64.exe //RS/neo4j). No spaces in the pathname will still be returned by the WQL filter in the detection script because there are spaces in the overall pathname but no quotes around the file path. This regex addition should better match and provide quotes only around the pathname but not the arguments.